bulk insert done for invoice payment,supplierinvoice and supplierinvoice pament with testing of all endpoints

+++ server/api/supplierinvoicepayment/sql/supplierinvoicepayment.initialize.sql
@@ -10,7 +10,7 @@ BEGIN
     [Amount] float NOT NULL,
     [Booked] bit NULL,
     [Currency] nvarchar(3) NULL,
-    [CurrencyRate] float NULL,
+    [CurrencyRate] decimal NULL,
     [CurrencyUnit] float NULL,
     [InvoiceNumber] int NOT NULL,
     [Number] int NULL,
